create tokens endpoint and move admin schemas
