+++ app/endpoints/admin/roles.py
@@ -5,7 +5,7 @@ from fastapi.responses import JSONResponse, Response
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.helpers._accesscontroller import AccessController
-from app.schemas.auth import PermissionType, Role, RoleRequest, Roles, RolesResponse, RoleUpdateRequest
+from app.schemas.admin.roles import PermissionType, Role, RoleRequest, Roles, RolesResponse, RoleUpdateRequest
 from app.sql.session import get_db_session
 from app.utils.configuration import configuration
 from app.utils.context import global_context, request_context
