+++ app/utils/variables.py
@@ -21,6 +21,7 @@ ENDPOINT__OCR = "/ocr-beta"
 ENDPOINT__PARSE = "/parse-beta"
 ENDPOINT__RERANK = "/rerank"
 ENDPOINT__SEARCH = "/search"
+ENDPOINT__TOKENS = "/tokens"
 ENDPOINT__USAGE = "/usage"
 
 ENDPOINTS = [value for name, value in locals().items() if name.startswith("ENDPOINT__")]
@@ -33,19 +34,20 @@ ROUTER__CHAT = "chat"
 ROUTER__CHUNKS = "chunks"
 ROUTER__COLLECTIONS = "collections"
 ROUTER__COMPLETIONS = "completions"
+ROUTER__DEEPSEARCH = "deepsearch"
 ROUTER__DOCUMENTS = "documents"
 ROUTER__EMBEDDINGS = "embeddings"
 ROUTER__FILES = "files"
 ROUTER__MODELS = "models"
 ROUTER__MONITORING = "monitoring"
+ROUTER__MULTIAGENTS = "multiagents"
+ROUTER__OAUTH2 = "oauth2"
 ROUTER__OCR = "ocr"
 ROUTER__PARSE = "parse"
 ROUTER__RERANK = "rerank"
 ROUTER__SEARCH = "search"
 ROUTER__USAGE = "usage"
-ROUTER__MULTIAGENTS = "multiagents"
-ROUTER__OAUTH2 = "oauth2"
-ROUTER__DEEPSEARCH = "deepsearch"
+ROUTER__TOKENS = "tokens"
 
 
 ROUTERS = [value for name, value in locals().items() if name.startswith("ROUTER__")]
